Remastered heredoc, quote, dquote and bquote to check juste before parsing, transform line with quote.. and backslash to something readable

+++ src/completion/deal_completion.c
@@ -6,7 +6,7 @@
 /*   By: rfontain <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/22 01:38:48 by rfontain          #+#    #+#             */
-/*   Updated: 2018/12/11 17:46:15 by rfontain         ###   ########.fr       */
+/*   Updated: 2018/12/12 16:56:42 by rfontain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 static void	deal_complet(t_tree *file, t_line *line)
 {
 	int		put;
-	int		j;
+	size_t	j;
 
 	tputs(tgetstr("sc", NULL), 1, ft_pchar);
 	tputs(tgetstr("do", NULL), 1, ft_pchar);
